working-app-4/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.40
openai>=1.0.0
pyyaml>=6.0
pydantic>=2.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
langchain-community>=0.1.0
langchain-ollama>=0.1.0